Task-Management/requirements.txt
asgiref==3.8.1
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.1.0
click==8.1.7
distlib==0.3.8
dj-database-url==2.3.0
Django==4.2.23
djangorestframework==3.15.2
drf-yasg==1.21.8
exceptiongroup==1.1.1
filelock==3.16.0
flake8==7.1.1
gunicorn==23.0.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
mccabe==0.7.0
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
mysqlclient==2.2.5
packaging==23.1
pathspec==0.12.1
platformdirs==4.3.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
pytest==8.3.3
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.4
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.5.0
virtualenv==20.31.2
python-decouple==3.8
djangorestframework-simplejwt~=5.5.0
pip~=24.3.1
attrs~=24.2.0
wheel~=0.37.0
PySocks~=1.7.1
protobuf~=6.30.2
multidict~=6.4.3
pygments~=2.19.2
setuptools~=44.1.1
future~=0.18.2
PyJWT~=2.9.0
django-cors-headers